… client/src/__tests__/components/EnhancedEntrantsGrid.test.tsx
… client/src/components/race-view/EnhancedEntrantsGrid.tsx
… client/src/components/race-view/MoneyFlowColumns.tsx
… client/src/components/race-view/PoolToggle.tsx
… client/src/components/race-view/RaceDataHeader.tsx
… client/src/components/race-view/RaceFooter.tsx
… client/src/components/race-view/RaceNavigation.tsx
… client/src/components/race-view/RacePageContent.tsx
… client/src/components/race-view/SortableColumns.tsx
… client/src/hooks/useEnhancedRealtime.ts
… client/src/hooks/useOptimizedRealtime.ts
… client/src/lib/cache.ts
… client/src/services/websocket.ts
… client/src/types/enhancedGrid.ts
… client/src/types/jockeySilks.ts
… client/src/types/meetings.ts
… client/src/types/moneyFlow.ts
… client/src/types/racePools.ts
… client/src/utils/accessibility.ts
… client/src/utils/performance.ts
… docs/architecture.md
… docs/architecture/1-system-overview.md
… docs/architecture/10-client-integration.md
… docs/architecture/11-conclusion.md
… docs/architecture/2-microservices-backend-architecture.md
… docs/architecture/3-frontend-architecture-nextjs-15.md
… docs/architecture/4-database-schema-appwrite.md
… docs/architecture/5-key-architectural-improvements.md
… docs/architecture/6-deployment-architecture.md
… docs/architecture/7-performance-characteristics.md